Renamed logos correctly
